@@ -24,7 +24,7 @@
     "test": "gulp test"
   },
   "peerDependencies": {
-    "react": "^15.3.0"
+    "react": "^15.3.0 || ^16.0.0"
   },
   "devDependencies": {
     "babel-core": "^6.13.2",
@@ -35,7 +35,8 @@
     "babel-preset-stage-2": "^6.5.0",
     "browser-sync": "^2.18.6",
     "chai": "^4.0.0",
-    "enzyme": "^2.7.1",
+    "enzyme": "^3.0.0",
+    "enzyme-adapter-react-16": "^1.0.0",
     "eslint": "^4.3.0",
     "eslint-config-takiyon-react": "^0.2.1",
     "eslint-import-resolver-webpack": "^0.8.1",
@@ -51,10 +52,9 @@
     "gulp-scss-lint": "^0.5.0",
     "jsdom": "^11.0.0",
     "mocha": "^3.0.2",
-    "react": "^15.4.2",
-    "react-addons-test-utils": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-test-renderer": "^15.5.4",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "react-test-renderer": "^16.0.0",
     "webpack": "^3.0.0",
     "webpack-stream": "^4.0.0"
   },
