@@ -1,4 +1,8 @@
-const { JSDOM } = require('jsdom');
+import Enzyme from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
+import { JSDOM } from 'jsdom';
+
+Enzyme.configure({ adapter: new Adapter() });
 
 const jsdom = new JSDOM('<!DOCTYPE html><html><body></body></html>');
 const { window } = jsdom;
